Initial commit
Adapted the NHEntity class in quokka (https://github.com/software-projects/quokka).
This is a useful class, but Quokka.Core has a number of other dependencies that
prevent its broader use.

Changed the behaviour of transient entities. In Quokka, changing the Id will cause
GetHashCode to throw an exception. Now, if GetHashCode is called on a transient
object

Updated the tests to provide 100% coverage.